Conflicts:
	inspire/Views/Experience/View/ExperienceDetails.swift
	inspire/Views/Explorer/Components/ExperienceCard.swift
	inspire/Views/Explorer/Components/ExperienceList.swift
	inspire/Views/Explorer/Views/ExplorerView.swift
	inspire/Views/Explorer/Views/MapView.swift
	inspire/Views/Favoris/Views/FavorisView.swift
	inspire/Views/Quiz/Components/Suggestion.swift
	inspire/Views/Quiz/View/Flow.swift
	inspire/Views/Quiz/View/MasterView.swift
	inspire/Views/Quiz/View/Suggestions.swift
	inspire/inspireApp.swift

+++ inspire/Components/PriceView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct PriceView: View {
-    let price : Double
+    let price : Int
     let color : Color
     var body: some View {
         HStack {
@@ -18,7 +18,7 @@ struct PriceView: View {
                 .foregroundColor(color)
         }
     }
-    func priceFormatter(_ price : Double) -> String {
+    func priceFormatter(_ price : Int) -> String {
         let formatter = NumberFormatter()
         formatter.locale = Locale.init(identifier: "fr")
         formatter.numberStyle = .currency
